fix(wi-1): CI — merge orphan test_inbox_durability into test__server, drop unused tempfile
Lead's CI heads-up (2026-05-20) on PR #126:
  1. ruff: ``tests/.../a2a/test_inbox_durability.py:25 import tempfile``
     is unused.
  2. audit-all (PS-204): the file is an orphan-test-file — no
     matching src/.../a2a/inbox_durability.py mirror.

The durability tests exercise behaviour in ``a2a/_server.py``
(``_publish_channel_event`` persists; ``get_inbox_stream`` replays +
stamps SSE ``id:`` lines), so the correct host is
``test__server.py``. Merged the four end-to-end tests into a new
"WI-1 — channel-bus durability + replay-on-reconnect" section at
the bottom of ``test__server.py``. The multi-assertion
``test_message_send_persists_event_to_channel_events`` is split into
five per-behaviour tests (TQ007 — one assertion per test) sharing a
single ``_durable_publish_row`` fixture:
  * test_publish_with_no_subscriber_persists_exactly_one_row
  * test_persisted_row_target_matches_path_param
  * test_persisted_row_source_carries_from_agent_metadata
  * test_persisted_row_content_carries_message_text
  * test_persisted_row_delivered_at_is_null_with_no_subscriber

Helpers (``_send_payload``, ``_run_loopback``, ``_free_port``,
``_consume_first_event``, ``_consume_event_with_id``) moved over;
the imports use underscore-prefixed aliases (``_asyncio``,
``_uvicorn``, …) to avoid colliding with the existing module-level
import block. ``_write_yaml`` was already defined at the top of
``test__server.py`` — reused as-is.

Deleted: tests/.../a2a/test_inbox_durability.py (no longer needed).

Tests:
  * pytest tests/.../a2a/test__server.py -k "publish_with_no_subscriber or persisted_row or event_posted or replayed_event or sse_id"
    — 8 passed.
  * pytest tests/.../a2a/ — 154 passed (no regressions).
  * Local audit gate: ``test_inbox_durability.py`` no longer in the
    PS-204 list. Remaining PS-204 entries are pre-existing on
    develop (test_host_config_resolve.py etc.) and out of scope per
    the lead's note that pre-existing develop issues are deferred
    to separate cleanup.
